Intorduce shared manager in snowpark
